+++ breeze-reflect/src/test/kotlin/com/windea/breezeframework/game/ClassExtensionsKtTest.kt
@@ -11,31 +11,31 @@ class ClassExtensionsKtTest {
 		assertTrue(String::class.java.isCharSequence)
 		assertTrue(CharSequence::class.java.isCharSequence)
 	}
-	
+
 	@Test
 	fun isString() {
 		assertTrue(String::class.java.isString)
 	}
-	
+
 	@Test
 	fun isIterable() {
 		assertTrue(List::class.java.isIterable)
 		assertTrue(ArrayList::class.java.isIterable)
 		assertTrue(Iterable::class.java.isIterable)
 	}
-	
+
 	@Test
 	fun isList() {
 		assertTrue(ArrayList::class.java.isList)
 		assertTrue(List::class.java.isList)
 	}
-	
+
 	@Test
 	fun isSet() {
 		assertTrue(HashSet::class.java.isSet)
 		assertTrue(Set::class.java.isSet)
 	}
-	
+
 	@Test
 	fun isMap() {
 		println(Map::class.java) //java.util.Map
@@ -48,25 +48,25 @@ class ClassExtensionsKtTest {
 		assertTrue(Map::class.java.isMap)
 		assertTrue(HashMap::class.java.isMap)
 	}
-	
+
 	@Test //TESTED
 	fun isSerializable() {
 		assertTrue(Foo::class.java.isSerializable)
 		assertTrue(Serializable::class.java.isSerializable)
 		//assertTrue(List::class.java.isSerializable) // will assert fail
 	}
-	
+
 	class Foo : Serializable {
 		val a = 1
 		var b = 2
 	}
-	
+
 	@Test //TESTED
 	fun getSetterMap() {
 		println(Foo::class.java.setterMap)
 		assertTrue(Foo::class.java.setterMap.size == 1)
 	}
-	
+
 	@Test //TESTED
 	fun getGetterMap() {
 		println(Foo::class.java.getterMap)