+++ CHANGELOG.md
@@ -234,6 +234,7 @@
 * [X] 将`DslBuilder`重命名为`DslDocument`，其他一些更改。
 * [X] 添加必要的Jvm注解，加强Java兼容性，其他一些相关更改。
 * [X] breeze-core 添加一些类型别名。
+* [X] breeze-core 移除自定义的`Cloneable<T>`接口，添加不稳定的`shallowClone`和`deepClone`扩展。移除`List.withKeys`扩展。
 
 # 待办